ref: move NotParsableUrlException to common
