+++ elasticms-cli/src/Client/Audit/AuditResult.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace App\CLI\Client\Audit;
 
 use App\CLI\Client\HttpClient\UrlReport;
-use App\CLI\Client\WebToElasticms\Helper\NotParsableUrlException;
 use App\CLI\Helper\HtmlHelper;
 use App\CLI\Helper\StringStream;
 use EMS\CommonBundle\Contracts\CoreApi\Endpoint\File\FileInterface;
+use EMS\CommonBundle\Exception\NotParsableUrlException;
 use EMS\CommonBundle\Helper\Url;
 
 class AuditResult
