+++ elasticms-cli/src/Client/WebToElasticms/Filter/Html/InternalLink.php
@@ -6,8 +6,8 @@ namespace App\CLI\Client\WebToElasticms\Filter\Html;
 
 use App\CLI\Client\WebToElasticms\Config\ConfigManager;
 use App\CLI\Client\WebToElasticms\Config\WebResource;
-use App\CLI\Client\WebToElasticms\Helper\NotParsableUrlException;
 use App\CLI\Client\WebToElasticms\Rapport\Rapport;
+use EMS\CommonBundle\Exception\NotParsableUrlException;
 use EMS\CommonBundle\Helper\Url;
 use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\Exception\RequestException;
